pyproject.toml
[tool.poetry]
name = "takehome"
version = "0.1.0"
description = ""
authors = ["Team Simplify <software@simplify.jobs>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.0"}
pydantic = "^2.9.1"
pydantic-settings = "^2.4.0"
httpx = "^0.27.2"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.35"
requests = "^2.32.2"
redis = "^5.0.8"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"


[tool.poetry.scripts]
mock = "scripts:mock"
dev = "scripts:dev"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

